Extract learn viewsets into viewsets/ package; delete viewsets.py
