--- kolibri/plugins/learn/viewsets.py
@@ -1,574 +1,0 @@
-from django.db.models import Count
-from django.db.models import OuterRef
-from django.db.models import Q
-from django.db.models import Subquery
-from django.db.models import Sum
-from django.db.models.fields import IntegerField
-from le_utils.constants import content_kinds
-from le_utils.constants import modalities
-from rest_framework import serializers
-from rest_framework.decorators import action
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.views import APIView
-
-from kolibri.core.api import ReadOnlyValuesViewset
-from kolibri.core.auth.models import Classroom
-from kolibri.core.auth.models import Facility
-from kolibri.core.content.api import ContentNodeProgressViewset
-from kolibri.core.content.api import ContentNodeViewset
-from kolibri.core.content.api import UserContentNodeViewset
-from kolibri.core.content.models import ContentNode
-from kolibri.core.courses.models import CourseSession
-from kolibri.core.exams.models import Exam
-from kolibri.core.exams.models import exam_assignment_lookup
-from kolibri.core.lessons.models import Lesson
-from kolibri.core.lessons.viewsets.lesson import ClassroomSerializer
-from kolibri.core.logger.models import AttemptLog
-from kolibri.core.logger.models import ContentSummaryLog
-from kolibri.core.logger.models import MasteryLog
-
-contentnode_progress_viewset = ContentNodeProgressViewset()
-contentnode_viewset = ContentNodeViewset()
-user_contentnode_viewset = UserContentNodeViewset()
-
-
-class LearnStateView(APIView):
-    def get(self, request, format=None):
-        """
-        Returns some configuration variables applicable to users navigating learn.
-        - in_classes: Whether the user is in any classes
-        - can_download_externally: Whether the user can download content externally from Kolibri
-        """
-        if request.user.is_anonymous:
-            default_facility = Facility.get_default_facility()
-            can_download_externally = (
-                default_facility.dataset.show_download_button_in_learn
-                if default_facility
-                else True
-            )
-            return Response(
-                {
-                    "in_classes": False,
-                    "can_download_externally": can_download_externally,
-                }
-            )
-        return Response(
-            {
-                "in_classes": request.user.memberships.exists(),
-                "can_download_externally": request.user.dataset.show_download_button_in_learn,
-            }
-        )
-
-
-def _map_contentnodes(request, content_ids):
-    contentnodes = (
-        contentnode_viewset.serialize_list(request, {"ids": content_ids})
-        if content_ids
-        else []
-    )
-    contentnode_map = {c["id"]: c for c in contentnodes}
-    return contentnode_map
-
-
-def _consolidate_lessons_data(request, lessons):
-    lesson_contentnode_ids = set()
-    for lesson in lessons:
-        lesson_contentnode_ids |= {
-            resource["contentnode_id"] for resource in lesson["resources"]
-        }
-
-    contentnode_progress = (
-        contentnode_progress_viewset.serialize_list(
-            request, {"ids": lesson_contentnode_ids}
-        )
-        if lesson_contentnode_ids
-        else []
-    )
-
-    contentnode_map = _map_contentnodes(request, lesson_contentnode_ids)
-
-    progress_map = {l["content_id"]: l["progress"] for l in contentnode_progress}
-
-    for lesson in lessons:
-        lesson["progress"] = {
-            "resource_progress": sum(
-                (
-                    progress_map[resource["content_id"]]
-                    for resource in lesson["resources"]
-                    if resource["content_id"] in progress_map
-                )
-            ),
-            "total_resources": len(lesson["resources"]),
-        }
-        missing_resource = False
-        for resource in lesson["resources"]:
-            resource["progress"] = progress_map.get(resource["content_id"], 0)
-            resource["contentnode"] = contentnode_map.get(
-                resource["contentnode_id"], None
-            )
-            missing_resource = missing_resource or not resource["contentnode"]
-        lesson["missing_resource"] = missing_resource
-
-
-def _consolidate_courses_data(request, courses, *, course_key="course"):
-    if not courses:
-        return courses
-
-    course_content_ids = {course[course_key] for course in courses}
-
-    course_nodes = ContentNode.objects.filter(id__in=course_content_ids).annotate(
-        unit_count=Count(
-            "children", filter=Q(children__modality=modalities.UNIT), distinct=True
-        ),
-        lesson_count=Count(
-            "children__children",
-            filter=Q(
-                children__modality=modalities.UNIT,
-                children__children__modality=modalities.LESSON,
-            ),
-            distinct=True,
-        ),
-    )
-
-    # Gather each course's non-topic descendant content ids in one pass, then
-    # look up the user's completed content across all courses in a single query
-    # rather than running one ContentSummaryLog count per course.
-    course_meta = {}
-    all_content_ids = set()
-    for course_node in course_nodes:
-        descendant_ids = list(
-            course_node.get_descendants()
-            .exclude(kind=content_kinds.TOPIC)
-            .values_list("content_id", flat=True)
-        )
-        course_meta[course_node.id] = (
-            course_node.unit_count,
-            course_node.lesson_count,
-            descendant_ids,
-        )
-        all_content_ids.update(descendant_ids)
-
-    completed_content_ids = set(
-        ContentSummaryLog.objects.filter(
-            user=request.user, progress=1, content_id__in=all_content_ids
-        ).values_list("content_id", flat=True)
-    )
-
-    course_data_map = {}
-    for course_id, (unit_count, lesson_count, descendant_ids) in course_meta.items():
-        total_content = len(descendant_ids)
-        if total_content:
-            user_completed_content = len(set(descendant_ids) & completed_content_ids)
-            progress = user_completed_content / total_content
-        else:
-            progress = 0
-
-        course_data_map[course_id] = {
-            "unit_count": unit_count,
-            "lesson_count": lesson_count,
-            "progress": progress,
-        }
-
-    for course in courses:
-        course_content_id = course[course_key]
-        data = course_data_map.get(
-            course_content_id, {"unit_count": 0, "lesson_count": 0, "progress": 0}
-        )
-        course.update(data)
-        if course_key == "course":
-            course["course_id"] = course.pop("course")
-
-    return courses
-
-
-class ExamProgressSerializer(serializers.Serializer):
-    closed = serializers.BooleanField(allow_null=True)
-    score = serializers.IntegerField(allow_null=True)
-    answer_count = serializers.IntegerField(allow_null=True)
-    started = serializers.BooleanField()
-
-
-class ExamLearnerSerializer(serializers.Serializer):
-    collection = serializers.CharField()
-    active = serializers.BooleanField()
-    archive = serializers.BooleanField()
-    id = serializers.CharField()
-    question_count = serializers.IntegerField()
-    title = serializers.CharField()
-    data_model_version = serializers.IntegerField()
-    question_sources = serializers.ListField()
-    instant_report_visibility = serializers.BooleanField(allow_null=True)
-    progress = ExamProgressSerializer()
-    missing_resource = serializers.BooleanField()
-
-
-class LessonProgressSerializer(serializers.Serializer):
-    resource_progress = serializers.FloatField()
-    total_resources = serializers.IntegerField()
-
-
-class LessonResourceSerializer(serializers.Serializer):
-    content_id = serializers.CharField()
-    channel_id = serializers.CharField()
-    contentnode_id = serializers.CharField()
-    progress = serializers.FloatField()
-    contentnode = serializers.DictField(allow_null=True)
-
-
-class LessonLearnerSerializer(serializers.Serializer):
-    description = serializers.CharField(allow_blank=True)
-    id = serializers.CharField()
-    active = serializers.BooleanField()
-    title = serializers.CharField()
-    resources = LessonResourceSerializer(many=True, read_only=True)
-    collection = serializers.CharField()
-    progress = LessonProgressSerializer()
-    missing_resource = serializers.BooleanField()
-
-
-class CourseLearnerSerializer(serializers.Serializer):
-    id = serializers.CharField()
-    course_id = serializers.CharField()
-    title = serializers.CharField()
-    description = serializers.CharField(allow_null=True, allow_blank=True)
-    is_active = serializers.BooleanField()
-    collection = serializers.CharField()
-    unit_count = serializers.IntegerField()
-    lesson_count = serializers.IntegerField()
-    progress = serializers.FloatField()
-
-
-class LearnerClassroomSerializer(serializers.ModelSerializer):
-    exams = ExamLearnerSerializer(many=True, read_only=True)
-    lessons = LessonLearnerSerializer(many=True, read_only=True)
-    courses = CourseLearnerSerializer(many=True, read_only=True)
-
-    class Meta:
-        model = Classroom
-        fields = ("id", "name", "exams", "lessons", "courses")
-
-
-class LearnerClassroomViewset(ReadOnlyValuesViewset):
-    """
-    Returns all Classrooms for which the requesting User is a member,
-    along with all associated assignments.
-    """
-
-    permission_classes = (IsAuthenticated,)
-    serializer_class = LearnerClassroomSerializer
-    deferred_fields = ("exams", "lessons", "courses")
-
-    def get_queryset(self):
-        if self.request.user.is_anonymous:
-            return Classroom.objects.none()
-        return Classroom.objects.filter(membership__user=self.request.user)
-
-    def consolidate(self, items, queryset):
-        if not items:
-            return items
-        lessons = (
-            Lesson.objects.filter(
-                lesson_assignments__collection__membership__user=self.request.user,
-                is_active=True,
-                collection__in=(c["id"] for c in items),
-            )
-            .distinct()
-            .values(
-                "description", "id", "is_active", "title", "resources", "collection"
-            )
-        )
-        _consolidate_lessons_data(self.request, lessons)
-        for lesson in lessons:
-            lesson["active"] = lesson.pop("is_active")
-
-        user_masterylog_content_ids = MasteryLog.objects.filter(
-            user=self.request.user
-        ).values("summarylog__content_id")
-
-        exams = (
-            Exam.objects.filter(
-                assignments__collection__membership__user=self.request.user,
-                collection__in=(c["id"] for c in items),
-            )
-            .filter(Q(active=True) | Q(id__in=user_masterylog_content_ids))
-            .annotate(
-                closed=Subquery(
-                    MasteryLog.objects.filter(
-                        summarylog__content_id=OuterRef("id"), user=self.request.user
-                    ).values("complete")[:1]
-                ),
-                score=Subquery(
-                    AttemptLog.objects.filter(
-                        sessionlog__content_id=OuterRef("id"), user=self.request.user
-                    )
-                    .order_by()
-                    .values_list("item")
-                    .distinct()
-                    .values("masterylog")
-                    .annotate(total_correct=Sum("correct"))
-                    .values("total_correct"),
-                    output_field=IntegerField(),
-                ),
-                answer_count=Subquery(
-                    AttemptLog.objects.filter(
-                        sessionlog__content_id=OuterRef("id"), user=self.request.user
-                    )
-                    .order_by()
-                    .values_list("item")
-                    .distinct()
-                    .values("masterylog")
-                    .annotate(total_complete=Count("id"))
-                    .values("total_complete"),
-                    output_field=IntegerField(),
-                ),
-            )
-            .distinct()
-            .values(
-                "collection",
-                "active",
-                "archive",
-                "id",
-                "question_count",
-                "title",
-                "closed",
-                "answer_count",
-                "data_model_version",
-                "score",
-                "question_sources",
-                "instant_report_visibility",
-            )
-        )
-        exam_node_ids = set()
-
-        for exam in exams:
-            exam_node_ids |= {
-                exercise_id
-                for exercise_id, _ in exam_assignment_lookup(
-                    exam.get("question_sources", [])
-                )
-            }
-
-        available_exam_ids = set(
-            ContentNode.objects.filter_by_uuids(exam_node_ids).values_list(
-                "id", flat=True
-            )
-        )
-
-        contentnode_map = _map_contentnodes(self.request, available_exam_ids)
-
-        for exam in exams:
-            closed = exam.pop("closed")
-            score = exam.pop("score")
-            answer_count = exam.pop("answer_count")
-            if closed is not None:
-                exam["progress"] = {
-                    "closed": closed,
-                    "score": score,
-                    "answer_count": answer_count,
-                    "started": True,
-                }
-            else:
-                exam["progress"] = {
-                    "score": None,
-                    "answer_count": None,
-                    "closed": None,
-                    "started": False,
-                }
-            missing_resource = False
-            for exercise_id, _ in exam_assignment_lookup(
-                exam.get("question_sources", [])
-            ):
-                if exercise_id not in contentnode_map:
-                    missing_resource = True
-                    break
-            exam["missing_resource"] = missing_resource
-
-        courses = (
-            CourseSession.objects.filter(
-                assignments__collection__membership__user=self.request.user,
-                collection__in=(c["id"] for c in items),
-                is_active=True,
-            )
-            .distinct()
-            .values("id", "course", "title", "description", "is_active", "collection")
-        )
-
-        courses = _consolidate_courses_data(self.request, courses) if courses else []
-
-        out_items = []
-        for item in items:
-            item["exams"] = [exam for exam in exams if exam["collection"] == item["id"]]
-            item["lessons"] = [
-                lesson for lesson in lessons if lesson["collection"] == item["id"]
-            ]
-            item["courses"] = [
-                course for course in courses if course["collection"] == item["id"]
-            ]
-            out_items.append(item)
-        return out_items
-
-
-learner_classroom_viewset = LearnerClassroomViewset()
-
-
-def _resumable_resources(classrooms):
-    for classroom in classrooms:
-        for lesson in classroom["lessons"]:
-            for resource in lesson["resources"]:
-                yield 0 < resource["progress"] < 1
-
-
-class LearnHomePageHydrationView(APIView):
-    def get(self, request, format=None):
-        classrooms = []
-        courses = []
-        resumable_resources = []
-        resumable_resources_progress = []
-        if not request.user.is_anonymous:
-            classrooms = learner_classroom_viewset.serialize_list(request)
-            for classroom in classrooms:
-                courses.extend(classroom.get("courses", []))
-            # Show out-of-course resumable content only when no classroom lesson
-            # resource is actively in-progress (0 < progress < 1); suppressed when
-            # the user has in-progress classroom work to focus on (b6d46e2822).
-            if not classrooms or not any(_resumable_resources(classrooms)):
-                resumable_resources = user_contentnode_viewset.serialize_list(
-                    request,
-                    {
-                        "resume": True,
-                        "max_results": 12,
-                        "ordering": "-last_interacted",
-                        "exclude_course_ancestry": True,
-                    },
-                )
-                resumable_resources_progress = (
-                    contentnode_progress_viewset.serialize_list(
-                        request,
-                        {
-                            "resume": True,
-                            "max_results": 12,
-                            "ordering": "-last_interacted",
-                            "exclude_course_ancestry": True,
-                        },
-                    )
-                )
-
-        return Response(
-            {
-                "classrooms": classrooms,
-                "courses": courses,
-                "resumable_resources": resumable_resources,
-                "resumable_resources_progress": resumable_resources_progress,
-            }
-        )
-
-
-class LearnerLessonSerializer(serializers.ModelSerializer):
-    classroom = ClassroomSerializer(source="collection", read_only=True)
-    active = serializers.BooleanField(source="is_active")
-    progress = serializers.DictField(read_only=True)
-    missing_resource = serializers.BooleanField(read_only=True)
-
-    class Meta:
-        model = Lesson
-        fields = (
-            "id",
-            "title",
-            "description",
-            "resources",
-            "active",
-            "collection",
-            "classroom",
-            "progress",
-            "missing_resource",
-        )
-
-
-class LearnerLessonViewset(ReadOnlyValuesViewset):
-    """
-    Special Viewset for Learners to view Lessons to which they are assigned.
-    The core Lesson Viewset is locked down to Admin users only.
-    """
-
-    permission_classes = (IsAuthenticated,)
-    serializer_class = LearnerLessonSerializer
-    deferred_fields = ("progress", "missing_resource")
-
-    def get_queryset(self):
-        if self.request.user.is_anonymous:
-            return Lesson.objects.none()
-        return Lesson.objects.filter(
-            lesson_assignments__collection__membership__user=self.request.user,
-            is_active=True,
-        ).select_related("collection")
-
-    def consolidate(self, items, queryset):
-        if not items:
-            return items
-
-        _consolidate_lessons_data(self.request, items)
-
-        return items
-
-
-class LearnerCourseClassroomSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Classroom
-        fields = ("id", "name", "parent")
-
-
-class LearnerCourseSerializer(serializers.ModelSerializer):
-    classroom = LearnerCourseClassroomSerializer(source="collection", read_only=True)
-    # The serializer renames "course" → "course_id" so the field_map produces
-    # "course_id" before consolidate(); _consolidate_courses_data is called with
-    # course_key="course_id" to match.
-    course_id = serializers.CharField(source="course")
-    unit_count = serializers.IntegerField(read_only=True)
-    lesson_count = serializers.IntegerField(read_only=True)
-    progress = serializers.FloatField(read_only=True)
-
-    class Meta:
-        model = CourseSession
-        fields = (
-            "id",
-            "course_id",
-            "title",
-            "description",
-            "is_active",
-            "classroom",
-            "unit_count",
-            "lesson_count",
-            "progress",
-        )
-
-
-class LearnerCourseViewset(ReadOnlyValuesViewset):
-    """
-    Special Viewset for Learners to view Course Sessions to which they are assigned.
-    """
-
-    permission_classes = (IsAuthenticated,)
-    serializer_class = LearnerCourseSerializer
-    deferred_fields = ("unit_count", "lesson_count", "progress")
-
-    def get_queryset(self):
-        if self.request.user.is_anonymous:
-            return CourseSession.objects.none()
-        # distinct() needed: a course session can be assigned to multiple
-        # sub-collections the user belongs to (e.g. classroom + learner group),
-        # which would produce duplicate rows without it (a4244f1b8e).
-        return CourseSession.objects.filter(
-            assignments__collection__membership__user=self.request.user,
-            is_active=True,
-        ).distinct()
-
-    def consolidate(self, items, queryset):
-        # The serializer field_map already renamed "course" → "course_id";
-        # pass course_key so _consolidate_courses_data reads the right key.
-        return _consolidate_courses_data(self.request, items, course_key="course_id")
-
-    @action(detail=True, methods=["get"])
-    def resume(self, request, pk=None):
-        course_session = self.get_object()
-        return Response(course_session.get_resume_data(request.user))
